Moved functional opts for JWT auth and auth method code to own files from utils.go
